full docstrings for geb/hazards/floods + move and rename some files
