+++ geb/hazards/floods/sfincs.py
@@ -1,736 +1,969 @@
+"""This module contains classes and functions to build and run SFINCS models for flood hazard assessment.
+
+The main class is `SFINCSRootModel`, which is used to create and manage SFINCS models.
+It provides methods to build the model, set up simulations with different forcing methods,
+and read simulation results.
+
+"""
+
 import json
-from collections import deque
+import shutil
+from datetime import datetime
 from pathlib import Path
+from typing import Any
 
 import geopandas as gpd
-import matplotlib.pyplot as plt
 import numpy as np
+import numpy.typing as npt
 import pandas as pd
-import rasterio
 import xarray as xr
-import zarr
-from shapely.geometry import shape
-from shapely.geometry.point import Point
+from hydromt_sfincs import SfincsModel
+from hydromt_sfincs.workflows import burn_river_rect, river_source_points
+from tqdm import tqdm
 
-from ...hydrology.HRUs import load_geom
-from ...hydrology.landcover import OPEN_WATER, SEALED
-from ...workflows.io import open_zarr, to_zarr
-from ...workflows.raster import reclassify
-from .model import SFINCSRootModel
+from geb.hydrology.routing import get_river_width
+
+from .workflows import do_mask_flood_plains, get_river_depth, get_river_manning
+from .workflows.return_periods import (
+    assign_calculation_group,
+    get_topological_stream_order,
+)
+from .workflows.utils import (
+    assign_return_periods,
+    create_hourly_hydrograph,
+    export_rivers,
+    get_discharge_and_river_parameters_by_river,
+    get_representative_river_points,
+    get_start_point,
+    import_rivers,
+    make_relative_paths,
+    read_flood_depth,
+    run_sfincs_simulation,
+    to_sfincs_datetime,
+)
 
 
-class SFINCS:
-    """The class that implements all methods to setup, run, and post-process the SFINCS hydrodynamic model.
+class SFINCSRootModel:
+    """Builds and updates SFINCS model files for flood hazard modeling."""
 
-    Args:
-        model: The GEB model instance.
-        n_timesteps: The number of timesteps to keep in memory for discharge calculations (default is 10).
+    def __init__(self, model: "GEBModel", event_name: str) -> None:
+        """Initializes the SFINCSRootModel with a GEBModel and event name.
+
+        Sets up the constant parts of the model (grid, mask, rivers, etc.),
+        and has methods to create simulations with actual forcing.
+
+        Args:
+            model: An instance of GEBModel containing hydrological and geographical data.
+            event_name: A string representing the name of the event (e.g., "flood_event_2023").
+                Also used to create the path to write the file to disk.
+        """
+        self.model = model
+        self.event_name = event_name
+
+    @property
+    def path(self) -> Path:
+        """Gets the root directory for the SFINCS model files.
+
+        Returns:
+            The path to the SFINCS model root directory.
+        """
+
+        folder: Path = self.model.simulation_root / "SFINCS" / self.event_name
+        folder.mkdir(parents=True, exist_ok=True)
+        return folder
+
+    def exists(self) -> bool:
+        """Checks if the SFINCS model already exists in the model root directory.
+
+        Returns:
+            True if the SFINCS model exists, False otherwise.
+        """
+        return Path(self.path / "sfincs.inp").is_file()
+
+    def read(self) -> None:
+        """Reads an existing SFINCS model from the model root directory.
+
+        Raises:
+            FileNotFoundError: if the SFINCS model does not exist in the specified path.
+        """
+        if not self.exists():
+            raise FileNotFoundError(f"SFINCS model not found in {self.path}")
+        self.sfincs_model = SfincsModel(root=str(self.path), mode="r")
+
+    def build(
+        self,
+        DEMs: list[dict[str, str]],
+        region: gpd.GeoDataFrame,
+        rivers: gpd.GeoDataFrame,
+        discharge: xr.DataArray,
+        waterbody_ids: npt.NDArray[np.int32],
+        river_width_alpha: npt.NDArray[np.float32],
+        river_width_beta: npt.NDArray[np.float32],
+        mannings: xr.DataArray,
+        resolution: float | int,
+        nr_subgrid_pixels: int | None,
+        crs: str,
+        depth_calculation_method: str,
+        depth_calculation_parameters: dict[str, float | int] | None = None,
+        mask_flood_plains: bool = False,
+    ) -> None:
+        """Build a SFINCS model.
+
+        Notes:
+            mask_flood_plains is currently quite unstable and should be used with caution. Sometimes it leads
+            to wrong regions being masked, which can lead to errors in the model.
+
+        Args:
+            DEMs: List of DEM datasets to use for the model. Should be a list of dictionaries with 'path' and 'name' keys.
+            region: A GeoDataFrame defining the region of interest.
+            rivers: A GeoDataFrame containing river segments.
+            discharge: An xarray DataArray containing discharge values for the rivers in m^3/s.
+            waterbody_ids: An numpy array of waterbody IDs specifying lakes and reservoirs. Should have same x and y dimensions as the discharge.
+            river_width_alpha: An numpy array of river width alpha parameters. Used for calculating river width.
+            river_width_beta: An numpy array of river width beta parameters. Used for calculating river width
+            mannings: A xarray DataArray of Manning's n values for the rivers.
+            resolution: The resolution of the requested SFINCS model grid in meters.
+            nr_subgrid_pixels: The number of subgrid pixels to use for the SFINCS model. Must be an even number.
+            crs: The coordinate reference system to use for the model.
+            depth_calculation_method: The method to use for calculating river depth. Can be 'manning' or 'power_law'.
+            depth_calculation_parameters: A dictionary of parameters for the depth calculation method. Only used if
+                depth_calculation_method is 'power_law', in which case it should contain 'c' and 'd' keys.
+            mask_flood_plains: Whether to autodelineate flood plains and mask them. Defaults to False.
+
+        Raises:
+            ValueError: if depth_calculation_method is not 'manning' or 'power_law',
+            ValueError: if nr_subgrid_pixels is not None and not positive even number.
+            ValueError: if resolution is not positive.
+
+        """
+        if nr_subgrid_pixels is not None and nr_subgrid_pixels <= 0:
+            raise ValueError("nr_subgrid_pixels must be a positive number")
+        if nr_subgrid_pixels is not None and nr_subgrid_pixels % 2 != 0:
+            raise ValueError("nr_subgrid_pixels must be an even number")
+        if resolution <= 0:
+            raise ValueError("Resolution must be a positive number")
+
+        assert depth_calculation_method in [
+            "manning",
+            "power_law",
+        ], "Method should be 'manning' or 'power_law'"
+
+        # build base model
+        sf: SfincsModel = SfincsModel(root=str(self.path), mode="w+")
+
+        sf.setup_grid_from_region(
+            {"geom": region}, res=resolution, crs=crs, rotated=False
+        )
+
+        DEMs = [{**DEM, **{"reproj_method": "bilinear"}} for DEM in DEMs]
+
+        # HydroMT-SFINCS only accepts datasets with an 'elevtn' variable. Therefore, the following
+        # is a bit convoluted. We first open the dataarray, then convert it to a dataset,
+        # and set the name as elevtn.
+        sf.setup_dep(datasets_dep=DEMs)
+
+        if mask_flood_plains:
+            do_mask_flood_plains(sf)
+        else:
+            sf.setup_mask_active(
+                region, zmin=-21, reset_mask=True
+            )  # TODO: Improve mask setup
+
+        # Setup river inflow points
+        sf.setup_river_inflow(
+            rivers=rivers,
+            keep_rivers_geom=True,
+            river_upa=0,
+            river_len=0,
+        )
+
+        # find outflow points and save for later use
+        outflow_points = river_source_points(
+            gdf_riv=rivers.to_crs(sf.crs),
+            gdf_mask=sf.region,
+            src_type="outflow",
+            buffer=sf.reggrid.dx,  # type: ignore
+            river_upa=0,
+            river_len=0,
+        )
+        # give error if outflow greater than 1
+        if len(outflow_points) > 1:
+            raise ValueError(
+                "More than one outflow point found, outflow boundary condition will fail to setup"
+            )
+        elif len(outflow_points) == 0:
+            raise ValueError(
+                "No outflow point found, outflow boundary condition will fail to setup"
+            )
+        # print crs of outflow_points
+        assert outflow_points.crs == sf.crs, (
+            "CRS of outflow_points is not the same as the model crs"
+        )
+        # set crs before saving
+        outflow_points = outflow_points.set_crs(sf.crs)
+        # save to model root as a gpkg file
+        outflow_points.to_file(self.path / "gis/outflow_points.gpkg", driver="GPKG")
+        # Get the single outflow point coordinates
+        x_coord = outflow_points.geometry.x.iloc[0]
+        y_coord = outflow_points.geometry.y.iloc[0]
+        assert sf.grid.dep.rio.crs == outflow_points.crs, (
+            "CRS of sf.grid.dep is not the same as the outflow_points crs"
+        )
+        # Sample from sf.grid.dep (which is the DEM DataArray)
+        elevation_value = sf.grid.dep.sel(
+            x=x_coord, y=y_coord, method="nearest"
+        ).values.item()
+
+        # Optional: sanity check
+        if elevation_value is None or elevation_value <= 0:
+            raise ValueError(
+                f"Invalid outflow elevation ({elevation_value}), must be > 0"
+            )
+
+        # Save elevation value to a file in model_root/gis
+        outflow_elev_path = self.path / "gis" / "outflow_elevation.json"
+        with open(outflow_elev_path, "w") as f:
+            json.dump({"outflow_elevation": elevation_value}, f)
+
+        river_representative_points = []
+        for ID in rivers.index:
+            river_representative_points.append(
+                get_representative_river_points(ID, rivers, waterbody_ids)
+            )
+
+        discharge_by_river, river_parameters = (
+            get_discharge_and_river_parameters_by_river(
+                rivers.index.tolist(),
+                river_representative_points,
+                discharge=discharge,
+                river_width_alpha=river_width_alpha,
+                river_width_beta=river_width_beta,
+            )
+        )
+
+        rivers = assign_return_periods(rivers, discharge_by_river, return_periods=[2])
+
+        river_width_unknown_mask = rivers["width"].isnull()
+
+        rivers.loc[river_width_unknown_mask, "width"] = get_river_width(
+            river_parameters["river_width_alpha"][river_width_unknown_mask],
+            river_parameters["river_width_beta"][river_width_unknown_mask],
+            rivers.loc[river_width_unknown_mask, "Q_2"],
+        )
+
+        rivers["depth"] = get_river_depth(
+            rivers,
+            method=depth_calculation_method,
+            parameters=depth_calculation_parameters,
+            bankfull_column="Q_2",
+        )
+
+        rivers["manning"] = get_river_manning(rivers)
+
+        export_rivers(self.path, rivers)
+
+        # Because hydromt-sfincs does a lot of filling default values when data
+        # is missing, we need to be extra sure that the required columns are
+        # present and contain valid data.
+        assert rivers["width"].notnull().all(), "River width cannot be null"
+        assert rivers["depth"].notnull().all(), "River depth cannot be null"
+        assert rivers["manning"].notnull().all(), "River Manning's n cannot be null"
+
+        # if sfincs is run with subgrid, we set up the subgrid, with burned in rivers and mannings
+        # roughness within the subgrid. If not, we burn the rivers directly into the main grid,
+        # including mannings roughness.
+        if nr_subgrid_pixels is not None:
+            print("Setting up SFINCS subgrid...")
+            sf.setup_subgrid(
+                datasets_dep=DEMs,
+                datasets_rgh=[
+                    {
+                        "manning": mannings.to_dataset(name="manning"),
+                    }
+                ],
+                datasets_riv=[
+                    {
+                        "centerlines": rivers.rename(
+                            columns={"width": "rivwth", "depth": "rivdph"}
+                        )
+                    }
+                ],
+                write_dep_tif=True,
+                write_man_tif=True,
+                nr_subgrid_pixels=nr_subgrid_pixels,
+                nlevels=20,
+                nrmax=500,
+            )
+
+            sf.write_subgrid()
+        else:
+            print("Setting up SFINCS without subgrid")
+            # first set up the mannings roughness with the default method
+            # (we already have the DEM set up)
+            sf.setup_manning_roughness(
+                datasets_rgh=[
+                    {
+                        "manning": mannings.to_dataset(name="manning"),
+                    }
+                ]
+            )
+            # retrieve the elevation and mannings grids
+            # burn the rivers into these grids
+            elevation, mannings = burn_river_rect(
+                da_elv=sf.grid.dep,
+                gdf_riv=rivers,
+                da_man=sf.grid.manning,
+                rivwth_name="width",
+                rivdph_name="depth",
+                manning_name="manning",
+                segment_length=sf.reggrid.dx,
+            )
+            # set the modified grids back to the model
+            sf.set_grid(elevation, name="dep")
+            sf.set_grid(mannings, name="manning")
+
+        # write all components, except forcing which must be done after the model building
+        sf.write_grid()
+        sf.write_geoms()
+        sf.write_config()
+        sf.write()
+
+        sf.plot_basemap(fn_out="basemap.png")
+
+        self.sfincs_model = sf
+
+    @property
+    def area(self) -> float | int:
+        """Returns the area of the SFINCS model region in square kilometers.
+
+        Returns:
+            The area of the SFINCS model region in m².
+        """
+        return self.sfincs_model.grid["msk"].sum().item() * self.cell_area
+
+    @property
+    def cell_area(self) -> float | int:
+        """Returns the area of a single cell in the SFINCS model grid in square meters.
+
+        Returns:
+            The area of a single cell in the SFINCS model grid in m².
+        """
+        return (
+            self.sfincs_model.grid["msk"].rio.resolution()[0]
+            * self.sfincs_model.grid["msk"].rio.resolution()[1]
+        )
+
+    def estimate_discharge_for_return_periods(
+        self,
+        discharge: xr.DataArray,
+        waterbody_ids: npt.NDArray[np.int32],
+        rivers: gpd.GeoDataFrame,
+        rising_limb_hours: int | float = 72,
+        return_periods: list[int | float] = [2, 5, 10, 20, 50, 100, 250, 500, 1000],
+    ) -> None:
+        """Estimate discharge for specified return periods and create hydrographs.
+
+        Args:
+            model_root: path to the SFINC model root directory
+            discharge: xr.DataArray containing the discharge data
+            waterbody_ids: array of waterbody IDs, of identical x and y dimensions as discharge
+            rivers: GeoDataFrame containing river segments
+            rising_limb_hours: number of hours for the rising limb of the hydrograph.
+            return_periods: list of return periods for which to estimate discharge.
+        """
+        recession_limb_hours = rising_limb_hours
+
+        # here we only select the rivers that have an upstream forcing point
+        rivers_with_forcing_point = rivers[~rivers["is_downstream_outflow_subbasin"]]
+
+        river_representative_points = []
+        for ID in rivers_with_forcing_point.index:
+            river_representative_points.append(
+                get_representative_river_points(
+                    ID, rivers_with_forcing_point, waterbody_ids
+                )
+            )
+
+        discharge_by_river, _ = get_discharge_and_river_parameters_by_river(
+            rivers_with_forcing_point.index,
+            river_representative_points,
+            discharge=discharge,
+        )
+        rivers_with_forcing_point = assign_return_periods(
+            rivers_with_forcing_point, discharge_by_river, return_periods=return_periods
+        )
+
+        for return_period in return_periods:
+            rivers_with_forcing_point[f"hydrograph_{return_period}"] = None
+
+        for river_idx in rivers_with_forcing_point.index:
+            for return_period in return_periods:
+                discharge_for_return_period = rivers_with_forcing_point.at[
+                    river_idx, f"Q_{return_period}"
+                ]
+                hydrograph = create_hourly_hydrograph(
+                    discharge_for_return_period,
+                    rising_limb_hours,
+                    recession_limb_hours,
+                )
+                hydrograph = {
+                    time.isoformat(): Q.item() for time, Q in hydrograph.iterrows()
+                }
+                rivers_with_forcing_point.at[
+                    river_idx, f"hydrograph_{return_period}"
+                ] = hydrograph
+
+        export_rivers(self.path, rivers_with_forcing_point, postfix="_return_periods")
+
+    def create_simulation(
+        self,
+        *args: Any,
+        **kwargs: Any,
+    ) -> "SFINCSSimulation":
+        """Sets forcing for a SFINCS model based on the provided parameters.
+
+        Creates a new simulation directory and creteas a new sfincs model
+        in that folder. Variables that do not change between simulations
+        (e.g., grid, mask, etc.) are retained in the original model folder
+        and inherited by the new simulation using relative paths.
+
+        Args:
+            *args: Positional arguments to pass to the SFINCSSimulation constructor.
+            **kwargs: Keyword arguments to pass to the SFINCSSimulation constructor.
+
+        Returns:
+            An instance of SFINCSSimulation with the configured forcing.
+        """
+        return SFINCSSimulation(
+            self,
+            *args,
+            **kwargs,
+        )
+
+    def create_simulation_for_return_period(
+        self, return_period: int | float
+    ) -> "MultipleSFINCSSimulations":
+        """Creates multiple SFINCS simulations for a specified return period.
+
+
+        The method groups rivers by their calculation group and creates a separate
+        simulation for each group. Each simulation is configured with discharge
+        hydrographs corresponding to the specified return period.
+        Args:
+            return_period: The return period for which to create simulations.
+
+        Returns:
+            An instance of MultipleSFINCSSimulations containing the created simulations.
+                This class aims to emulate a single SFINCSSimulation instance as if
+                it was one.
+        """
+        rivers: gpd.GeoDataFrame = import_rivers(self.path, postfix="_return_periods")
+        assert (~rivers["is_downstream_outflow_subbasin"]).all()
+
+        rivers["topological_stream_order"] = get_topological_stream_order(rivers)
+        rivers: gpd.GeoDataFrame = assign_calculation_group(rivers)
+
+        working_dir: Path = self.path / "working_dir"
+
+        print(f"Running SFINCS for return period {return_period} years")
+        simulations = []
+
+        working_dir_return_period: Path = working_dir / f"rp_{return_period}"
+        for group, group_rivers in tqdm(rivers.groupby("calculation_group")):
+            simulation_root = working_dir_return_period / str(group)
+
+            shutil.rmtree(simulation_root, ignore_errors=True)
+            simulation_root.mkdir(parents=True, exist_ok=True)
+
+            inflow_nodes = group_rivers.copy()
+            inflow_nodes = inflow_nodes.reset_index(drop=True)
+            inflow_nodes["geometry"] = inflow_nodes["geometry"].apply(get_start_point)
+
+            Q: list[pd.DataFrame] = [
+                pd.DataFrame.from_dict(
+                    inflow_nodes[f"hydrograph_{return_period}"].iloc[idx],
+                    orient="index",
+                    columns=[idx],
+                )
+                for idx in inflow_nodes.index
+            ]
+            Q: pd.DataFrame = pd.concat(Q, axis=1)
+            Q.index = pd.to_datetime(Q.index)
+
+            Q: pd.DataFrame = (
+                Q.fillna(method="ffill").fillna(  # pad with 0's before
+                    method="bfill"
+                )  # and after
+            )
+
+            simulation = self.create_simulation(
+                simulation_name=f"rp_{return_period}_group_{group}",
+                start_time=Q.index[0],
+                end_time=Q.index[-1],
+            )
+
+            simulation.set_discharge_forcing_from_nodes(
+                nodes=inflow_nodes.to_crs(self.sfincs_model.crs),
+                timeseries=Q,
+            )
+            simulations.append(simulation)
+
+        return MultipleSFINCSSimulations(simulations)
+
+
+class MultipleSFINCSSimulations:
+    """Manages multiple SFINCS simulations as a single entity."""
+
+    def __init__(self, simulations: list["SFINCSSimulation"]) -> None:
+        """Simulates running multiple SFINCS simulations as one.
+
+        Args:
+            simulations: A list of SFINCSSimulation instances to manage together.
+        """
+        self.simulations = simulations
+
+    def run(self, gpu: bool) -> None:
+        """Runs all contained SFINCS simulations.
+
+        Args:
+            gpu: Whether to use GPU acceleration for the simulations.
+        """
+        for simulation in self.simulations:
+            simulation.run(gpu=gpu)
+
+    def read_max_flood_depth(self) -> xr.DataArray:
+        """Reads the maximum flood depth map from the simulation output.
+
+        Returns:
+            An xarray DataArray containing the maximum flood depth.
+        """
+        flood_depths: list[xr.DataArray] = []
+        for simulation in self.simulations:
+            flood_depths.append(simulation.read_max_flood_depth())
+
+        rp_map: xr.DataArray = xr.concat(flood_depths, dim="node")
+        rp_map: xr.DataArray = rp_map.max(dim="node")
+        rp_map.attrs["_FillValue"] = flood_depths[0].attrs["_FillValue"]
+        assert rp_map.rio.crs is not None
+
+        return rp_map
+
+    def cleanup(self) -> None:
+        """Cleans up all simulation directories."""
+        for simulation in self.simulations:
+            simulation.cleanup()
+
+
+class SFINCSSimulation:
+    """A SFINCS simulation with specific forcing and configuration.
+
+    Created fro m a SFINCSRootModel instance which already contains the constant parts of the model.
     """
 
-    def __init__(self, model, n_timesteps=10) -> None:
-        self.model = model
-        self.config = (
-            self.model.config["hazards"]["floods"]
-            if "floods" in self.model.config["hazards"]
-            else {}
-        )
-
-        self.HRU = model.hydrology.HRU
-
-        if self.model.simulate_hydrology:
-            self.hydrology = model.hydrology
-            self.n_timesteps = n_timesteps
-            self.discharge_per_timestep = deque(maxlen=self.n_timesteps)
-            self.soil_moisture_per_timestep = deque(maxlen=self.n_timesteps)
-            self.max_water_storage_per_timestep = deque(maxlen=self.n_timesteps)
-            self.saturated_hydraulic_conductivity_per_timestep = deque(
-                maxlen=self.n_timesteps
-            )
-            self.soil_storage_capacity_per_timestep = deque(maxlen=self.n_timesteps)
-
-    def sfincs_model_root(self, basin_id) -> Path:
-        folder: Path = self.model.simulation_root / "SFINCS" / str(basin_id)
-        folder.mkdir(parents=True, exist_ok=True)
-        return folder
-
-    def sfincs_simulation_root(self, event) -> Path:
-        name: str = self.get_event_name(event)
-        folder: Path = (
-            self.sfincs_model_root(name)
-            / "simulations"
-            / f"{event['start_time'].strftime('%Y%m%dT%H%M%S')} - {event['end_time'].strftime('%Y%m%dT%H%M%S')}"
-        )
-        if self.model.multiverse_name:
-            folder: Path = folder / self.model.multiverse_name
-        folder.mkdir(parents=True, exist_ok=True)
-        return folder
-
-    def get_event_name(self, event):
-        if "name" in event:
-            return event["name"]
-        elif "basin_id" in event:
-            return event["basin_id"]
-        else:
-            return "run"
-
-    def get_utm_zone(self, region_file: Path | str) -> str:
-        """Determine the UTM zone based on the centroid of the region geometry.
-
-        Args:
-            region_file: Path to the region geometry file.
-
-        Returns:
-            The EPSG code for the UTM zone of the centroid of the region.
-        """
-        region: gpd.GeoDataFrame = load_geom(region_file)
-
-        # Calculate the central longitude of the dataset
-        centroid: Point = region.union_all().centroid
-
-        # Determine the UTM zone based on the longitude
-        utm_zone: int = int((centroid.x + 180) // 6) + 1
-
-        # Determine if the data is in the Northern or Southern Hemisphere
-        # The EPSG code for UTM in the northern hemisphere is EPSG:326xx (xx = zone)
-        # The EPSG code for UTM in the southern hemisphere is EPSG:327xx (xx = zone)
-        if centroid.y > 0:
-            utm_crs: str = f"EPSG:326{utm_zone}"  # Northern hemisphere
-        else:
-            utm_crs: str = f"EPSG:327{utm_zone}"  # Southern hemisphere
-        return utm_crs
-
-    def build(self, name) -> SFINCSRootModel:
-        """Builds or reads a SFINCS model without any forcing.
-
-        Before using this model, forcing must be set.
-
-        When the model already exists and force_overwrite is False, the existing model is read.
-
-        Args:
-            name: Name of the SFINCS model (used for the model root directory).
-
-        Returns:
-            The built or read SFINCSRootModel instance.
-        """
-        sfincs_model = SFINCSRootModel(self.model, name)
-        if self.config["force_overwrite"] or not sfincs_model.exists():
-            with open(self.model.files["dict"]["hydrodynamics/DEM_config"]) as f:
-                DEM_config = json.load(f)
-                for entry in DEM_config:
-                    entry["elevtn"] = open_zarr(
-                        self.model.files["other"][entry["path"]]
-                    ).to_dataset(name="elevtn")
-
-            sfincs_model.build(
-                region=load_geom(self.model.files["geom"]["routing/subbasins"]),
-                DEMs=DEM_config,
-                rivers=self.rivers,
-                discharge=self.discharge_spinup_ds,
-                waterbody_ids=self.model.hydrology.grid.decompress(
-                    self.model.hydrology.grid.var.waterBodyID
-                ),
-                river_width_alpha=self.model.hydrology.grid.decompress(
-                    self.model.var.river_width_alpha
-                ),
-                river_width_beta=self.model.hydrology.grid.decompress(
-                    self.model.var.river_width_beta
-                ),
-                mannings=self.mannings,
-                resolution=self.config["resolution"],
-                nr_subgrid_pixels=self.config["nr_subgrid_pixels"],
-                crs=self.crs,
-                depth_calculation_method=self.model.config["hydrology"]["routing"][
-                    "river_depth"
-                ]["method"],
-                depth_calculation_parameters=self.model.config["hydrology"]["routing"][
-                    "river_depth"
-                ]["parameters"]
-                if "parameters"
-                in self.model.config["hydrology"]["routing"]["river_depth"]
-                else {},
-                mask_flood_plains=False,  # setting this to True sometimes leads to errors
-            )
-        else:
-            sfincs_model.read()
-
-        return sfincs_model
-
-    def set_forcing(
-        self, sfincs_model, event, start_time, precipitation_scale_factor=1.0
-    ) -> None:
-        n_timesteps: int = min(self.n_timesteps, len(self.discharge_per_timestep))
-        routing_substeps: int = self.discharge_per_timestep[0].shape[0]
-        discharge_grid = self.hydrology.grid.decompress(
-            np.vstack(self.discharge_per_timestep)
-        )
-        first_timestep_of_event = len(self.discharge_per_timestep) - n_timesteps  #
-
-        # when SFINCS starts with high values, this leads to numerical instabilities. Therefore, we first start with very low discharge and then build up slowly to timestep 0
-        # TODO: Check if this is a right approach
-        discharge_grid = np.vstack(
-            [
-                np.full_like(
-                    discharge_grid[:routing_substeps, :, :], fill_value=np.nan
-                ),
-                discharge_grid,
-            ]
-        )  # prepend zeros
-
-        for i in range(routing_substeps - 1, -1, -1):
-            discharge_grid[i] = discharge_grid[i + 1] * 0.3
-
-        # convert the discharge grid to an xarray DataArray
-        discharge_grid = xr.DataArray(
-            data=discharge_grid,
-            coords={
-                "time": pd.date_range(
-                    end=self.model.current_time
-                    + self.model.timestep_length
-                    - self.model.timestep_length / routing_substeps,
-                    periods=(n_timesteps + 1)
-                    * routing_substeps,  # +1 because we prepend the discharge above
-                    freq=self.model.timestep_length / routing_substeps,
-                    inclusive="right",
-                ),
-                "y": self.hydrology.grid.lat,
-                "x": self.hydrology.grid.lon,
-            },
-            dims=["time", "y", "x"],
-            name="discharge",
-        )
-
-        self.initial_soil_moisture_grid = xr.Dataset(
-            {
-                "initial_soil_moisture": (
-                    ["y", "x"],
-                    np.flipud(
-                        self.HRU.decompress(
-                            self.soil_moisture_per_timestep[first_timestep_of_event]
-                        )
-                    ),  # take first time step of soil moisture
-                )
-            },  # deque
-            coords={
-                "y": self.HRU.lat,
-                "x": self.HRU.lon,
-            },
-        )
-
-        self.max_water_storage_grid = xr.Dataset(
-            {
-                "max_water_storage": (
-                    ["y", "x"],
-                    self.HRU.decompress(
-                        self.max_water_storage_per_timestep[first_timestep_of_event],
-                    ),  # take first time step of soil moisture
-                )
-            },  # deque
-            coords={
-                "y": self.HRU.lat,
-                "x": self.HRU.lon,
-            },
-        )
-
-        soil_storage_capacity_grid = xr.Dataset(
-            {
-                "soil_storage_capacity": (
-                    ["y", "x"],
-                    self.HRU.decompress(
-                        self.soil_storage_capacity_per_timestep[first_timestep_of_event]
-                    ),  # take first time step of soil moisture
-                )
-            },  # deque
-            coords={
-                "y": self.HRU.lat,
-                "x": self.HRU.lon,
-            },
-        )
-
-        saturated_hydraulic_conductivity_grid = xr.Dataset(
-            {
-                "saturated_hydraulic_conductivity": (
-                    ["y", "x"],
-                    self.HRU.decompress(
-                        self.saturated_hydraulic_conductivity_per_timestep[
-                            first_timestep_of_event
-                        ]
-                    ),  # take first time step of soil moisture
-                )
-            },  # deque
-            coords={
-                "y": self.HRU.lat,
-                "x": self.HRU.lon,
-            },
-        )
-
-        self.initial_soil_moisture_grid.raster.set_crs(
-            self.model.crs
-        )  # for soil moisture
-        saturated_hydraulic_conductivity_grid.raster.set_crs(
-            self.model.crs
-        )  # for saturated hydraulic conductivity
-        self.max_water_storage_grid.raster.set_crs(
-            self.model.crs
-        )  # for max water storage
-        soil_storage_capacity_grid.raster.set_crs(
-            self.model.crs
-        )  # for the soil water storage capacity
-
-        discharge_grid.raster.set_crs(self.model.crs)
-        end_time = discharge_grid.time[-1] + pd.Timedelta(
-            self.model.timestep_length / routing_substeps
-        )
-        discharge_grid: xr.DataArray = discharge_grid.sel(
-            time=slice(start_time, end_time)
-        )
-
-        precipitation_grid: xr.DataArray = self.model.forcing["pr_hourly"]
-        assert isinstance(precipitation_grid, xr.DataArray)
-        precipitation_grid: xr.DataArray = (
-            precipitation_grid * precipitation_scale_factor
-        )
-
-        simulation = sfincs_model.create_simulation(
-            simulation_name=f"{event['start_time'].strftime('%Y%m%dT%H%M%S')} - {event['end_time'].strftime('%Y%m%dT%H%M%S')}",
-            start_time=event["start_time"],
-            end_time=event["end_time"],
-        )
-        if self.config["forcing_method"] == "headwater_points":
-            simulation.set_headwater_forcing_from_grid(
-                discharge_grid=discharge_grid,
-                waterbody_ids=self.model.hydrology.grid.decompress(
-                    self.model.hydrology.grid.var.waterBodyID
-                ),
-            )
-        elif self.config["forcing_method"] == "precipitation":
-            simulation.set_precipitation_forcing(
-                soil_water_capacity_grid=soil_storage_capacity_grid,
-                max_water_storage_grid=self.max_water_storage_grid,
-                saturated_hydraulic_conductivity_grid=saturated_hydraulic_conductivity_grid,
-                precipitation_grid=precipitation_grid,
-            )
-        else:
-            raise ValueError(
-                f"Unknown forcing method {self.config['forcing_method']}. Supported are 'headwater_points' and 'precipitation'."
-            )
-        return simulation
-
-    def run_single_event(self, event, start_time, precipitation_scale_factor=1.0):
-        sfincs_root_model = self.build("entire_region")
-        sfincs_simulation = sfincs_root_model.create_simulation()
-
-        sfincs_simulation = self.set_forcing(
-            sfincs_root_model, event, start_time, precipitation_scale_factor
-        )
-        self.model.logger.info(f"Running SFINCS for {self.model.current_time}...")
-
-        sfincs_simulation.run(
-            gpu=self.config["SFINCS"]["gpu"],
-        )
-        flood_depth = sfincs_simulation.read_flood_depth()
-
-        flood_depth: xr.DataArray = to_zarr(
-            flood_depth,
-            self.model.output_folder / "flood_maps" / sfincs_simulation.name,
-            crs=flood_depth.rio.crs,
-        )
-        self.model.agents.households.flood(flood_depth=flood_depth)
-
-    def build_mask_for_coastal_sfincs(self) -> gpd.GeoDataFrame:
-        """Builds a mask to define the active cells and boundaries for the coastal SFINCS model.
-
-        Returns:
-            GeoDataFrame: A GeoDataFrame containing the coastal mask.
-        """
-        # Load the dataset (assumes NetCDF with CF conventions and georeferencing info)
-        mask = xr.load_dataset(self.model.files["other"]["drainage/mask"])
-
-        # Extract the mask variable
-        mask_var = mask["mask"]
-
-        # Make sure it has a CRS
-        if mask_var.rio.crs is None:
-            mask_var = mask_var.rio.write_crs(
-                "EPSG:4326", inplace=False
-            )  # or your known CRS
-
-        # Extract binary mask values
-        mask_data = mask_var.values.astype(np.uint8)
-
-        # Get transform from raster metadata
-        transform = mask_var.rio.transform()
-
-        # Use rasterio.features.shapes() to get polygons for each contiguous region with same value
-        shapes = rasterio.features.shapes(mask_data, mask=None, transform=transform)
-
-        # Build GeoDataFrame from the shapes generator
-        records = [{"geometry": shape(geom), "value": value} for geom, value in shapes]
-
-        gdf = gpd.GeoDataFrame.from_records(records)
-        gdf.set_geometry("geometry", inplace=True)
-        gdf.crs = mask_var.rio.crs
-        # include a 1km buffer to the mask to include the coastal areas
-        # Keep only mask == 1
-        gdf = gdf[gdf["value"] == 1]
-        # gdf.geometry = gdf.geometry.buffer(0.00833)
-
-        return gdf
-
-    def build_coastal_boundary_mask(self) -> gpd.GeoDataFrame:
-        """Builds a mask to define the coastal boundaries for the SFINCS model.
-
-        Returns:
-            GeoDataFrame: A GeoDataFrame containing the coastal boundary mask.
-        """
-        lecz = xr.load_dataset(
-            self.model.files["other"]["landsurface/low_elevation_coastal_zone"]
-        )
-
-        # Make sure it has a CRS
-        if lecz.rio.crs is None:
-            lecz = lecz.rio.write_crs(
-                "EPSG:4326", inplace=False
-            )  # check CRS for later applications
-
-        # Extract binary mask values
-        lecz_data = lecz["low_elevation_coastal_zone"].values.astype(np.uint8)
-
-        # Get transform from raster metadata
-        transform = lecz.rio.transform()
-
-        # Use rasterio.features.shapes() to get polygons for each contiguous region with same value
-        shapes = rasterio.features.shapes(lecz_data, mask=None, transform=transform)
-
-        # Build GeoDataFrame from the shapes generator
-        records = [{"geometry": shape(geom), "value": value} for geom, value in shapes]
-
-        gdf = gpd.GeoDataFrame.from_records(records)
-        gdf.set_geometry("geometry", inplace=True)
-        gdf = gdf.set_crs(lecz.rio.crs, inplace=True)
-        gdf = gdf[gdf["value"] == 1]  # Keep only mask == 1
-        return gdf
-
-    def get_coastal_return_period_maps(self) -> dict[int, xr.DataArray]:
-        """This function models coastal flooding for the return periods specified in the model config.
-
-        Returns:
-            dict[int, xr.DataArray]: A dictionary mapping return periods to their respective flood maps.
-        """
-        coastal_mask = self.build_mask_for_coastal_sfincs()
-        boundary_mask = self.build_coastal_boundary_mask()
-        model_root: Path = self.sfincs_model_root("entire_region_coastal")
-        build_parameters = self.get_build_parameters(model_root)
-        build_parameters["region"] = coastal_mask
-        build_parameters["boundary_mask"] = boundary_mask
-        build_sfincs_coastal(
-            **build_parameters,
-        )
-
-        rp_maps_coastal = run_sfincs_for_return_periods_coastal(
-            model=self.model,
-            model_root=model_root,
-            gpu=self.config["SFINCS"]["gpu"],
-            export_dir=self.model.output_folder / "flood_maps",
-            clean_working_dir=True,
-            return_periods=self.config["return_periods"],
-        )
-        return rp_maps_coastal
-
-    def get_riverine_return_period_maps(self) -> dict[int, xr.DataArray]:
-        """This function models riverine flooding for the return periods specified in the model config.
-
-        Returns:
-            dict[int, xr.DataArray]: A dictionary mapping return periods to their respective flood maps.
-        """
-        # close the zarr store
-        if hasattr(self.model, "reporter"):
-            self.model.reporter.variables["discharge_daily"].close()
-
-        sfincs_root_model = self.build("entire_region")
-
-        sfincs_root_model.estimate_discharge_for_return_periods(
-            discharge=self.discharge_spinup_ds,
-            waterbody_ids=self.model.hydrology.grid.decompress(
-                self.model.hydrology.grid.var.waterBodyID
-            ),
-            rivers=self.rivers,
-            return_periods=self.config["return_periods"],
-        )
-
-        rp_maps_riverine = {}
-        for return_period in self.config["return_periods"]:
-            print(
-                f"Estimated discharge for return period {return_period} years for all rivers."
-            )
-
-            simulation = sfincs_root_model.create_simulation_for_return_period(
-                return_period,
-            )
-            simulation.run(
-                gpu=self.config["SFINCS"]["gpu"],
-            )
-            flood_depth_return_period = simulation.read_flood_depth()
-            rp_maps_riverine[return_period] = flood_depth_return_period
-
-        if hasattr(self.model, "reporter"):
-            # and re-open afterwards
-            self.model.reporter.variables["discharge_daily"] = zarr.ZipStore(
-                self.model.config["report_hydrology"]["discharge_daily"]["path"],
-                mode="a",
-            )
-
-        return rp_maps_riverine
-
-    def merge_return_period_maps(
+    def __init__(
         self,
-        rp_maps_coastal: dict[int, xr.DataArray],
-        rp_maps_riverine: dict[int, xr.DataArray],
+        sfincs_root_model: SFINCSRootModel,
+        simulation_name: str,
+        start_time: datetime,
+        end_time: datetime,
+        spinup_seconds: int = 86400,
+        write_gis_files: bool = True,
+        write_figures: bool = False,
     ) -> None:
-        """Merges the return period maps for riverine and coastal floods into a single dataset.
+        """Initializes a SFINCSSimulation with specific forcing and configuration.
 
         Args:
-            rp_maps_coastal: Dictionary of coastal return period maps.
-            rp_maps_riverine: Dictionary of riverine return period maps.
+            sfincs_root_model: An instance of SFINCSRootModel containing the base model.
+            simulation_name: A string representing the name of the simulation.
+                Also used to create the path to write the file to disk.
+            start_time: The start time of the simulation as a datetime object.
+            end_time: The end time of the simulation as a datetime object.
+            spinup_seconds: The number of seconds to use for model spin-up. Defaults to 86400 (1 day).
+            write_gis_files: Whether to write GIS files for the model. Defaults to True.
+            write_figures: Whether to generate and save figures for the model. Defaults to False.
         """
-        for return_period in self.config["return_periods"]:
-            if rp_maps_coastal is None:
-                to_zarr(
-                    da=rp_maps_riverine[return_period],
-                    path=self.model.output_folder
-                    / "flood_maps"
-                    / f"{return_period}.zarr",
-                    crs=rp_maps_riverine[return_period].rio.crs,
+        self._name = simulation_name
+        self.write_figures = write_figures
+        self.start_time = start_time
+        self.end_time = end_time
+        self.sfincs_root_model = sfincs_root_model
+
+        sfincs_model = sfincs_root_model.sfincs_model
+        sfincs_model.set_root(self.path, mode="w+")
+
+        # update mode time based on event tstart and tend from event dict
+        sfincs_model.setup_config(
+            alpha=0.5
+        )  # alpha is the parameter for the CFL-condition reduction. Decrease for additional numerical stability, minimum value is 0.1 and maximum is 0.75 (0.5 default value)
+        sfincs_model.setup_config(tspinup=spinup_seconds)  # spinup time in seconds
+        sfincs_model.setup_config(dtout=900)  # output time step in seconds
+        sfincs_model._write_gis = write_gis_files
+        sfincs_model.setup_config(
+            tref=to_sfincs_datetime(start_time),
+            tstart=to_sfincs_datetime(start_time),
+            tstop=to_sfincs_datetime(end_time),
+        )
+        sfincs_model.setup_config(
+            **make_relative_paths(sfincs_model.config, self.root_path, self.path)
+        )
+
+        sfincs_model.write_config()
+
+        self.sfincs_model = sfincs_model
+
+    def set_headwater_forcing_from_grid(
+        self,
+        discharge_grid: str | xr.DataArray,
+        waterbody_ids: npt.NDArray[np.int32],
+    ) -> None:
+        """Sets up discharge forcing for the SFINCS model from a gridded dataset.
+
+        Args:
+            discharge_grid: Path to a raster file or an xarray DataArray containing discharge values in m^3/s.
+                Usually this is from a hydrological model.
+            waterbody_ids: An numpy array of waterbody IDs specifying lakes and reservoirs.
+                Should have same x and y dimensions as the discharge.
+        """
+        rivers: gpd.GeoDataFrame = import_rivers(self.root_path)
+        rivers_with_forcing_point: gpd.GeoDataFrame = rivers[
+            ~rivers["is_downstream_outflow_subbasin"]
+        ]
+        headwater_rivers: gpd.GeoDataFrame = rivers_with_forcing_point[
+            rivers_with_forcing_point["maxup"] == 0
+        ]
+
+        inflow_nodes: gpd.GeoDataFrame = headwater_rivers.copy()
+
+        # Only select headwater points. Maxup is the number of upstream river segments.
+        inflow_nodes["geometry"] = inflow_nodes["geometry"].apply(get_start_point)
+
+        river_representative_points = []
+        for ID in headwater_rivers.index:
+            river_representative_points.append(
+                get_representative_river_points(ID, headwater_rivers, waterbody_ids)
+            )
+
+        discharge_by_river, _ = get_discharge_and_river_parameters_by_river(
+            headwater_rivers.index,
+            river_representative_points,
+            discharge=discharge_grid,
+        )
+
+        locations = inflow_nodes.to_crs(self.sfincs_model.crs)
+        index_mapping = {
+            idx: i + 1
+            for i, idx in enumerate(locations.index)  # SFINCS index starts at 1
+        }
+        locations.index = locations.index.map(index_mapping)
+        locations.index.name = "sfincs_idx"
+        discharge_by_river.columns = discharge_by_river.columns.map(index_mapping)
+
+        self.set_discharge_forcing_from_nodes(
+            nodes=locations,
+            timeseries=discharge_by_river,
+        )
+
+    def set_discharge_forcing_from_nodes(
+        self, nodes: gpd.GeoDataFrame, timeseries: pd.DataFrame
+    ) -> None:
+        """Sets up discharge forcing for the SFINCS model from specified nodes (locations) and timeseries.
+
+        Args:
+            nodes: A GeoDataFrame containing the locations of the discharge forcing points.
+                The index should start at 1 and be consecutive (1, 2, 3, ...).
+            timeseries: A DataFrame containing the discharge timeseries for each node.
+                The columns should match the index of the nodes GeoDataFrame.
+        """
+        assert np.array_equal(nodes.index, np.arange(1, len(nodes) + 1))
+        assert set(timeseries.columns) == set(nodes.index)
+
+        self.sfincs_model.setup_discharge_forcing(
+            locations=nodes,
+            timeseries=timeseries,
+        )
+
+        self.sfincs_model.write_forcing()
+        self.sfincs_model.write_config()
+
+        if self.write_figures:
+            self.sfincs_model.plot_forcing(fn_out="forcing.png")
+            self.sfincs_model.plot_basemap(fn_out="basemap.png")
+
+    def set_precipitation_forcing_grid(
+        self,
+        current_water_storage_grid: xr.DataArray,
+        max_water_storage_grid: xr.DataArray,
+        saturated_hydraulic_conductivity_grid: xr.DataArray,
+        precipitation_grid: xr.DataArray,
+    ) -> None:
+        """Sets up precipitation forcing for the SFINCS model from a gridded dataset.
+
+        Args:
+            current_water_storage_grid: xarray DataArray containing current soil water capacity [m].
+            max_water_storage_grid: xarray DataArray containing maximum water storage [m].
+            saturated_hydraulic_conductivity_grid: xarray DataArray containing saturated hydraulic conductivity [m/s].
+            precipitation_grid: xarray DataArray containing precipitation values in kg/m²/s (equivalent to mm/s).
+        """
+        assert precipitation_grid.raster.crs is not None, (
+            "precipitation_grid should have a crs"
+        )
+        assert (
+            pd.to_datetime(precipitation_grid.time[0].item()).to_pydatetime()
+            <= self.start_time
+        )
+        assert (
+            pd.to_datetime(precipitation_grid.time[-1].item()).to_pydatetime()
+            >= self.end_time
+        )
+
+        precipitation_grid: xr.DataArray = precipitation_grid.sel(
+            time=slice(self.start_time, self.end_time)
+        )
+
+        self.sfincs_model.setup_precip_forcing_from_grid(
+            precip=(precipitation_grid * 3600).to_dataset(name="precip")
+        )  # convert from kg/m2/s to mm/h
+
+        self._setup_infiltration_capacity(
+            max_water_storage=max_water_storage_grid,
+            current_water_storage=current_water_storage_grid,
+            saturated_hydraulic_conductivity=saturated_hydraulic_conductivity_grid,
+        )
+        self.sfincs_model.write_forcing()
+        self.sfincs_model.write_config()
+
+        if self.write_figures:
+            self.sfincs_model.plot_basemap(fn_out="basemap.png")
+            self.sfincs_model.plot_forcing(fn_out="forcing.png")
+
+    # def setup_outflow_boundary(self) -> None:
+    #     # detect whether water level forcing should be set (use this under forcing == coastal) PLot basemap and forcing to check
+    #     if (
+    #         self.sfincs_model.grid["msk"] == 2
+    #     ).any():  # if mask is 2, the model requires water level forcing
+    #         waterlevel = self.sfincs_model.data_catalog.get_dataset(
+    #             "waterlevel"
+    #         ).compute()  # define water levels and stations in data_catalog.yml
+
+    #         locations = gpd.GeoDataFrame(
+    #             index=waterlevel.stations,
+    #             geometry=gpd.points_from_xy(
+    #                 waterlevel.station_x_coordinate, waterlevel.station_y_coordinate
+    #             ),
+    #             crs=4326,
+    #         )
+
+    #         timeseries = pd.DataFrame(
+    #             index=waterlevel.time, columns=waterlevel.stations, data=waterlevel.data
+    #         )
+    #         assert timeseries.columns.equals(locations.index)
+
+    #         locations = locations.reset_index(names="stations")
+    #         locations.index = (
+    #             locations.index + 1
+    #         )  # for hydromt/SFINCS index should start at 1
+    #         timeseries.columns = locations.index
+
+    #         self.sfincs_model.setup_waterlevel_forcing(
+    #             timeseries=timeseries, locations=locations
+    #         )
+
+    #     self.sfincs_model.write_forcing()
+
+    def _setup_infiltration_capacity(
+        self,
+        max_water_storage: xr.DataArray,
+        current_water_storage: xr.DataArray,
+        saturated_hydraulic_conductivity: xr.DataArray,
+        initial_substractiion: float = 0.0,
+    ) -> None:
+        """Set up infiltration parameters in the SFINCS model.
+
+        Uses the curve number method with recovery.
+
+        Args:
+            sfincs_model: SfincsModel object to update.
+            max_water_storage: xarray DataArray containing maximum water storage [m].
+            current_water_storage: xarray DataArray containing soil water capacity [m].
+            saturated_hydraulic_conductivity: xarray DataArray containing saturated hydraulic conductivity [m/s].
+            initial_substractiion: Initial abstraction ratio [-].
+                Is removed from rainfall before infiltration. Defaults to 0.0.
+        """
+        remaining_water_storage = max_water_storage - current_water_storage
+        remaining_water_storage = remaining_water_storage.compute()
+
+        # maximum water storage (smax in SFINCS)
+        max_water_storage = max_water_storage.raster.reproject_like(
+            self.sfincs_model.grid, method="average"
+        )
+        assert not np.isnan(max_water_storage.values[self.active_cells]).any(), (
+            "max_water_storage contains NaN values in active cells"
+        )
+        max_water_storage = max_water_storage.to_dataset(name="smax")
+        self.sfincs_model.set_grid(max_water_storage, name="smax")
+        self.sfincs_model.set_config("smaxfile", "sfincs.smax")
+
+        # remaining water storage (seff in SFINCS)
+        remaining_water_storage = remaining_water_storage.raster.reproject_like(
+            self.sfincs_model.grid, method="nearest"
+        )
+        assert not np.isnan(remaining_water_storage.values[self.active_cells]).any(), (
+            "current_water_storage contains NaN values in active cells"
+        )
+        remaining_water_storage = remaining_water_storage.to_dataset(name="seff")
+        self.sfincs_model.set_grid(remaining_water_storage, name="seff")
+        self.sfincs_model.set_config("sefffile", "sfincs.seff")
+
+        # saturated hydraulic conductivity (ks in SFINCS)
+        saturated_hydraulic_conductivity = (
+            (
+                saturated_hydraulic_conductivity.raster.reproject_like(
+                    self.sfincs_model.grid, method="average"
                 )
-                continue
-
-            coastal_da = rp_maps_coastal[return_period]
-            riverine_da = rp_maps_riverine[return_period]
-
-            # --- 2. Get union bounds ---
-            riv_bounds = riverine_da.rio.bounds()  # (minx, miny, maxx, maxy)
-            coa_bounds = coastal_da.rio.bounds()
-
-            minx = min(riv_bounds[0], coa_bounds[0])
-            miny = min(riv_bounds[1], coa_bounds[1])
-            maxx = max(riv_bounds[2], coa_bounds[2])
-            maxy = max(riv_bounds[3], coa_bounds[3])
-
-            # --- 3. Pick resolution ---
-            # Use riverine resolution (y is negative if north-up, so take abs)
-            res_x, res_y = riverine_da.rio.resolution()
-            res_x = abs(res_x)
-            res_y = abs(res_y)
-
-            # --- 4. Build template coords ---
-            width = int(np.ceil((maxx - minx) / res_x))
-            height = int(np.ceil((maxy - miny) / res_y))
-
-            x_coords = minx + (np.arange(width) + 0.5) * res_x
-            y_coords = maxy - (np.arange(height) + 0.5) * res_y  # top→bottom
-
-            template = xr.DataArray(
-                np.full((height, width), np.nan, dtype=riverine_da.dtype),
-                coords={"y": y_coords, "x": x_coords},
-                dims=("y", "x"),
-            ).rio.write_crs(riverine_da.rio.crs)
-
-            # --- 5. Reproject both datasets to the template ---
-            riverine_reproj = riverine_da.rio.reproject_match(template)
-            coastal_reproj = coastal_da.rio.reproject_match(template)
-
-            # --- 6. Merge via maximum ---
-            rp_map = xr.concat([riverine_reproj, coastal_reproj], dim="stacked").max(
-                dim="stacked", skipna=True
             )
-            rp_map.rio.write_crs(riverine_da.rio.crs)
-            to_zarr(
-                da=rp_map,
-                path=self.model.output_folder / "flood_maps" / f"{return_period}.zarr",
-                crs=rp_map.rio.crs,
-            )
+            * 3600
+            * 1000
+        )  # convert from m/s to mm/h for SFINCS
+        assert not np.isnan(
+            saturated_hydraulic_conductivity.values[self.active_cells]
+        ).any(), "saturated_hydraulic_conductivity contains NaN values in active cells"
 
-    def run(self, event) -> None:
-        start_time = event["start_time"]
-
-        if self.model.config["hazards"]["floods"]["flood_risk"]:
-            scale_factors = pd.read_parquet(
-                self.model.files["table"]["hydrodynamics/risk_scaling_factors"]
-            )
-            scale_factors["return_period"] = 1 / scale_factors["exceedance_probability"]
-            damages_list = []
-            return_periods_list = []
-            exceedence_probabilities_list = []
-
-            for _, row in scale_factors.iterrows():
-                return_period = row["return_period"]
-                exceedence_probability = row["exceedance_probability"]
-
-                damages = self.run_single_event(
-                    event, start_time, precipitation_scale_factor=row["scaling_factor"]
-                )
-
-                damages_list.append(damages)
-                return_periods_list.append(return_period)
-                exceedence_probabilities_list.append(exceedence_probability)
-
-            print(damages_list)
-            print(return_periods_list)
-            print(exceedence_probabilities_list)
-
-            plt.plot(return_periods_list, damages_list)
-            plt.xlabel("Return period")
-            plt.ylabel("Flood damages [euro]")
-            plt.title("Damages per return period")
-            plt.show()
-
-            inverted_damage_list = damages_list[::-1]
-            inverted_exceedence_probabilities_list = exceedence_probabilities_list[::-1]
-
-            expected_annual_damage = np.trapz(
-                y=inverted_damage_list, x=inverted_exceedence_probabilities_list
-            )  # np.trapezoid or np.trapz -> depends on np version
-            print(f"exptected annual damage is: {expected_annual_damage}")
-
-        else:
-            self.run_single_event(event, start_time)
-
-    def save_discharge(self) -> None:
-        self.discharge_per_timestep.append(
-            self.hydrology.grid.var.discharge_m3_s_per_substep
-        )  # this is a deque, so it will automatically remove the oldest discharge
-
-    def save_soil_moisture(self) -> None:  # is used in driver.py on every timestep
-        # load and process initial soil moisture grid
-        w_copy = self.HRU.var.w.copy()
-        w_copy[:, self.HRU.var.land_use_type == SEALED] = 0
-        w_copy[:, self.HRU.var.land_use_type == OPEN_WATER] = 0
-        self.initial_soil_moisture_grid = w_copy[:2].sum(axis=0)
-        self.soil_moisture_per_timestep.append(self.initial_soil_moisture_grid)
-
-    def save_max_soil_moisture(self) -> None:
-        # smax
-        ws_copy = self.HRU.var.ws.copy()
-        ws_copy[:, self.HRU.var.land_use_type == SEALED] = 0
-        ws_copy[:, self.HRU.var.land_use_type == OPEN_WATER] = 0
-        self.max_water_storage_grid = ws_copy[:2].sum(axis=0)
-        self.max_water_storage_per_timestep.append(self.max_water_storage_grid)
-
-    def save_soil_storage_capacity(self) -> None:
-        self.soil_storage_capacity_grid = (
-            self.max_water_storage_grid - self.initial_soil_moisture_grid
+        saturated_hydraulic_conductivity: xr.Dataset = (
+            saturated_hydraulic_conductivity.to_dataset(name="ks")
         )
-        self.soil_storage_capacity_per_timestep.append(self.soil_storage_capacity_grid)
+        self.sfincs_model.set_grid(
+            saturated_hydraulic_conductivity, name="ks"
+        )  # convert from m/s to mm/h
+        self.sfincs_model.set_config("ksfile", "sfincs.ks")
 
-    def save_saturated_hydraulic_conductivity(self) -> None:
-        saturated_hydraulic_conductivity_copy = (
-            self.HRU.var.saturated_hydraulic_conductivity.copy()
-        )
-        saturated_hydraulic_conductivity_copy[
-            :, self.HRU.var.land_use_type == SEALED
-        ] = 0
-        saturated_hydraulic_conductivity_copy[
-            :, self.HRU.var.land_use_type == OPEN_WATER
-        ] = 0
-        saturated_hydraulic_conductivity_copy = saturated_hydraulic_conductivity_copy[
-            :2
-        ].sum(axis=0)
+        # initial abstraction ratio
+        assert initial_substractiion >= 0.0, "initial_substraction must be non-negative"
+        assert initial_substractiion < 1.0, "initial_substraction must be < 1.0"
 
-        saturated_hydraulic_conductivity_copy = (
-            saturated_hydraulic_conductivity_copy * 1000
-        ) / 24
-        self.saturated_hydraulic_conductivity_per_timestep.append(
-            saturated_hydraulic_conductivity_copy
+        self.sfincs_model.set_config("sfacinf", initial_substractiion)
+
+        # remove constant infiltration rate if present
+        self.sfincs_model.config.pop("qinf", None)
+
+        # write grids and config
+        self.sfincs_model.write_grid(data_vars=["smax", "seff", "ks"])
+        self.sfincs_model.write_config()
+
+    def run(self, gpu: bool) -> None:
+        """Runs the SFINCS simulation.
+
+        Args:
+            gpu: Whether to use GPU acceleration for the simulation.
+        """
+        run_sfincs_simulation(
+            simulation_root=self.path,
+            model_root=self.root_path,
+            gpu=gpu,
         )
 
-    @property
-    def discharge_spinup_ds(self) -> xr.DataArray:
-        """Open the discharge datasets from the model output folder."""
-        da: xr.DataArray = open_zarr(
-            self.model.output_folder
-            / "report"
-            / "spinup"
-            / "hydrology.routing"
-            / "discharge_daily.zarr"
-        )
+    def read_max_flood_depth(self, minimum_flood_depth: int | float) -> xr.DataArray:
+        """Reads the maximum flood depth map from the simulation output.
 
-        # start_time = pd.to_datetime(ds.time[0].item()) + pd.DateOffset(years=10)
-        # ds = ds.sel(time=slice(start_time, ds.time[-1]))
-
-        # # make sure there is at least 20 years of data
-        # if not len(ds.time.groupby(ds.time.dt.year).groups) >= 20:
-        #     raise ValueError(
-        #         """Not enough data available for reliable spinup, should be at least 20 years of data left.
-        #         Please run the model for at least 30 years (10 years of data is discarded)."""
-        #     )
-
-        return da
-
-    @property
-    def rivers(self):
-        """Load the river geometry from the model files.
+        Args:
+            minimum_flood_depth: Minimum flood depth to consider [m]. Values below this threshold are set to zero.
 
         Returns:
-            A GeoDataFrame containing the river geometry.
+            An xarray DataArray containing the maximum flood depth.
         """
-        return load_geom(self.model.files["geom"]["routing/rivers"])
-
-    @property
-    def mannings(self) -> xr.DataArray:
-        """Get the Manning's n values for the land cover types."""
-        mannings = reclassify(
-            self.land_cover,
-            self.land_cover_mannings_rougness_classification.set_index(
-                "esa_worldcover"
-            )["N"].to_dict(),
-            method="lookup",
+        flood_map: xr.DataArray = read_flood_depth(
+            model_root=self.root_path,
+            simulation_root=self.path,
+            method="max",
+            minimum_flood_depth=minimum_flood_depth,
         )
-        return mannings
+        return flood_map
 
-    @property
-    def land_cover(self) -> xr.DataArray:
-        """Get the land cover classification for the model.
+    def read_final_flood_depth(self, minimum_flood_depth: float | int) -> xr.DataArray:
+        """Reads the final flood depth map from the simulation output.
 
-        Returns:
-            An xarray DataArray containing the land cover classification.
-        """
-        return open_zarr(self.model.files["other"]["landcover/classification"])
-
-    @property
-    def land_cover_mannings_rougness_classification(self) -> pd.DataFrame:
-        """Get the land cover classification table for Manning's roughness.
+        Args:
+            minimum_flood_depth: Minimum flood depth to consider [m]. Values below this threshold are set to zero.
 
         Returns:
-            A DataFrame containing the land cover classification for Manning's roughness.
+            An xarray DataArray containing the final flood depth.
         """
-        return pd.DataFrame(
-            data=[
-                [10, "Tree cover", 10, 0.12],
-                [20, "Shrubland", 20, 0.05],
-                [30, "Grasland", 30, 0.034],
-                [40, "Cropland", 40, 0.037],
-                [50, "Built-up", 50, 0.1],
-                [60, "Bare / sparse vegetation", 60, 0.023],
-                [70, "Snow and Ice", 70, 0.01],
-                [80, "Permanent water bodies", 80, 0.02],
-                [90, "Herbaceous wetland", 90, 0.035],
-                [95, "Mangroves", 95, 0.07],
-                [100, "Moss and lichen", 100, 0.025],
-                [0, "No data", 0, 0.1],
-            ],
-            columns=["esa_worldcover", "description", "landuse", "N"],
+        flood_map: xr.DataArray = read_flood_depth(
+            model_root=self.root_path,
+            simulation_root=self.path,
+            method="final",
+            minimum_flood_depth=minimum_flood_depth,
         )
+        return flood_map
 
-    @property
-    def crs(self) -> str:
-        """Get the coordinate reference system (CRS) for the model.
+    def get_flood_volume(self, flood_depth: xr.DataArray) -> float:
+        """Compute the total flood volume from the flood depth map.
 
-        When the CRS is set in the configuration, it will return that value.
-        If the CRS is set to "auto", it will determine the UTM zone based on the routing subbasins geometry.
+        Args:
+            flood_depth: An xarray DataArray containing the flood depth.
 
         Returns:
-             The CRS string, either "auto" or the determined UTM zone.
+            The total flood volume in cubic meters.
         """
-        crs: str = self.config["crs"]
-        if crs == "auto":
-            crs: str = self.get_utm_zone(self.model.files["geom"]["routing/subbasins"])
-        return crs
+        pixel_area = abs(
+            flood_depth.rio.resolution()[0] * flood_depth.rio.resolution()[1]
+        )
+        flooded_pixels = flood_depth.where(flood_depth > 0).sum().item()
+        if hasattr(flooded_pixels, "compute"):
+            flooded_pixels = flooded_pixels.compute()
+        return flooded_pixels * pixel_area
+
+    def cleanup(self) -> None:
+        """Cleans up the simulation directory by removing temporary files."""
+        shutil.rmtree(self.path, ignore_errors=True)
+
+    @property
+    def root_path(self) -> Path:
+        """Returns the root directory for the SFINCS model files."""
+        return self.sfincs_root_model.path
+
+    @property
+    def path(self) -> Path:
+        """Returns the root directory for the SFINCS simulation files."""
+        folder: Path = self.root_path / "simulations" / self.name
+        folder.mkdir(parents=True, exist_ok=True)
+        return folder
+
+    @property
+    def name(self) -> str:
+        """Returns the name of the simulation."""
+        return self._name
+
+    def has_outflow_boundary(self) -> bool:
+        """Checks if the SFINCS model has an outflow boundary condition.
+
+        Returns:
+            True if the SFINCS model has an outflow boundary condition, False otherwise.
+        """
+        return (self.sfincs_model.grid["msk"] == 2).any()
+
+    @property
+    def active_cells(self) -> xr.DataArray:
+        """Returns a boolean mask of the active cells in the SFINCS model.
+
+        Returns:
+            A boolean mask of the active cells in the SFINCS model.
+        """
+        return self.sfincs_model.grid["msk"] == 1
+
+    def get_cumulative_precipitation(self) -> xr.DataArray:
+        """Reads the cumulative precipitation from the SFINCS model results.
+
+        Returns:
+            An xarray DataArray containing the cumulative precipitation.
+        """
+        self.sfincs_model.read_results()
+        return self.sfincs_model.results["cumprcp"].isel(timemax=-1)
+
+    def get_cumulative_infiltration(self) -> xr.DataArray:
+        """Reads the cumulative infiltration from the SFINCS model results.
+
+        Returns:
+            An xarray DataArray containing the cumulative infiltration.
+        """
+        self.sfincs_model.read_results()
+        return self.sfincs_model.results["cuminf"].isel(timemax=-1)
